Export hashInfo and relocate hashInfo module
Move builder/hashInfo.js into src/core/Torrent/hashInfo.js and update imports accordingly. Add a re-export in src/core/Torrent.js for hashInfo and HashInfoError so the API surface exposes these from the Torrent entry point. Update tests to import from the new path. No functional changes to the implementation.
