+++ src/core/Torrent.js
@@ -1,4 +1,5 @@
 import { init, $client } from "./Torrent/init.js"
+export { hashInfo, HashInfoError } from "./Torrent/hashInfo.js"
 import { on } from "./Torrent/on.js"
 import { off } from "./Torrent/off.js"
 import { add } from "./Torrent/add.js"
