+++ src/core/Torrent/hashInfo.js
